Mike | add separate pages for players
