+++ components/Player/Pages/index.js
@@ -1,1 +1,4 @@
 export { default as PeersPage } from './Peers';
+export { default as OverviewPage } from './Overview';
+export { default as MatchesPage } from './Matches';
+export { default as HeroesPage } from './Heroes';
