+++ components/Player/Player.css
@@ -1,33 +1,3 @@
-@import "../palette.css";
-
 .header {
   margin-bottom: 30px;
-}
-
-.overviewContainer {
-  display: flex;
-}
-
-.overviewMatches {
-  width: 75%;
-  display: inline-block;
-}
-
-.overviewHeroes {
-  width: 25%;
-  display: inline-block;
-}
-
-.heroesContainer {
-  margin-left: 30px;
-}
-
-.card {
-  background-color: var(--textSecondaryColor) !important;
-  box-shadow: rgba(0, 0, 0, 0.7) 0px 1px 6px, rgba(0, 0, 0, 0.7) 0px 1px 4px !important;
-}
-
-.tableHeading {
-  margin-bottom: 10px;
-  font-size: 20px;
 }
